Add db-client package.
