+++ packages/api/.env
@@ -1,5 +1,3 @@
-DATABASE_URL=
-
 EMAIL_FROM=noreply@example.com
 ORIGIN_ALLOWLIST=http://localhost:4200,http://localhost:4300
 API_ORIGIN=http://localhost:4300
