--- packages/api/.gitignore
@@ -1,1 +1,0 @@
-prisma/client
