+++ packages/api/shared/db.ts
@@ -1,13 +1,9 @@
-import { PrismaClient, Prisma } from '../prisma/client';
-import { randomBytes } from 'crypto';
+import {
+  Prisma,
+  PrismaTypes,
+  PrismaCasl,
+  client,
+  ulid,
+} from '@translation/db-client';
 
-export { Prisma };
-export * as PrismaTypes from '../prisma/client';
-
-export const client = new PrismaClient();
-
-export function ulid() {
-  const timestamp = Date.now().toString(16).padStart(12, '0');
-  const random = randomBytes(10).toString('hex');
-  return `${timestamp}${random}`;
-}
+export { Prisma, PrismaTypes, PrismaCasl, client, ulid };
