+++ packages/db-client/.env
@@ -1,0 +1,1 @@
+DATABASE_URL=
