+++ packages/db-client/.gitignore
@@ -1,0 +1,1 @@
+prisma/client
