+++ packages/db-client/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@translation/db-client",
+  "version": "0.0.1",
+  "type": "commonjs"
+}
